refactoring; gave up on system check because it runs after AppConfig.ready
